third_party/blink/web_tests/fast/frames/avoid-forced-layout-on-empty-subframe-expected.txt

Test that inserting an empty iframe doesn't force a style-recalc/layout

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.nodeNeedsStyleRecalc(div) is true
PASS internals.layoutCountForTesting() is layoutCountBefore
PASS internals.nodeNeedsStyleRecalc(div) is true
PASS internals.nodeNeedsStyleRecalc(iframe) is true
PASS successfullyParsed is true

TEST COMPLETE
